updating for np.float64

+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_assembly_peg_v2.py
@@ -32,11 +32,11 @@ class SawyerNutAssemblyEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             "obj_init_angle": 0.3,
-            "obj_init_pos": np.array([0, 0.6, 0.02], dtype=np.float32),
-            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float32),
+            "obj_init_pos": np.array([0, 0.6, 0.02], dtype=np.float64),
+            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float64),
         }
 
-        self.goal = np.array([0.1, 0.8, 0.1], dtype=np.float32)
+        self.goal = np.array([0.1, 0.8, 0.1], dtype=np.float64)
         self.obj_init_pos = self.init_config["obj_init_pos"]
         self.obj_init_angle = self.init_config["obj_init_angle"]
         self.hand_init_pos = self.init_config["hand_init_pos"]
@@ -46,8 +46,8 @@ class SawyerNutAssemblyEnvV2(SawyerXYZEnv):
             np.hstack((obj_high, goal_high)),
         )
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
