+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_basketball_v2.py
@@ -33,8 +33,8 @@ class SawyerBasketballEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             "obj_init_angle": 0.3,
-            "obj_init_pos": np.array([0, 0.6, 0.03], dtype=np.float32),
-            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float32),
+            "obj_init_pos": np.array([0, 0.6, 0.03], dtype=np.float64),
+            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float64),
         }
         self.goal = np.array([0, 0.9, 0])
         self.obj_init_pos = self.init_config["obj_init_pos"]
