+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_button_press_wall_v2.py
@@ -27,8 +27,8 @@ class SawyerButtonPressWallEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            "obj_init_pos": np.array([0.0, 0.9, 0.115], dtype=np.float32),
-            "hand_init_pos": np.array([0, 0.4, 0.2], dtype=np.float32),
+            "obj_init_pos": np.array([0.0, 0.9, 0.115], dtype=np.float64),
+            "hand_init_pos": np.array([0, 0.4, 0.2], dtype=np.float64),
         }
         self.goal = np.array([0, 0.84, 0.12])
         self.obj_init_pos = self.init_config["obj_init_pos"]
@@ -38,13 +38,13 @@ class SawyerButtonPressWallEnvV2(SawyerXYZEnv):
         goal_high = self.hand_high
 
         self._random_reset_space = Box(
-            np.array(obj_low, dtype=np.float32),
-            np.array(obj_high, dtype=np.float32),
+            np.array(obj_low, dtype=np.float64),
+            np.array(obj_high, dtype=np.float64),
         )
 
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
