+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_door_v2.py
@@ -43,12 +43,12 @@ class SawyerDoorEnvV2(SawyerXYZEnv):
         self.door_qvel_adr = self.model.joint("doorjoint").dofadr.item()
 
         self._random_reset_space = Box(
-            np.array(obj_low, dtype=np.float32),
-            np.array(obj_high, dtype=np.float32),
+            np.array(obj_low, dtype=np.float64),
+            np.array(obj_high, dtype=np.float64),
         )
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
