+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_plate_slide_back_v2.py
@@ -30,8 +30,8 @@ class SawyerPlateSlideBackEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             "obj_init_angle": 0.3,
-            "obj_init_pos": np.array([0.0, 0.85, 0.0], dtype=np.float32),
-            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float32),
+            "obj_init_pos": np.array([0.0, 0.85, 0.0], dtype=np.float64),
+            "hand_init_pos": np.array((0, 0.6, 0.2), dtype=np.float64),
         }
         self.goal = np.array([0.0, 0.6, 0.015])
         self.obj_init_pos = self.init_config["obj_init_pos"]
@@ -43,8 +43,8 @@ class SawyerPlateSlideBackEnvV2(SawyerXYZEnv):
             np.hstack((obj_high, goal_high)),
         )
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
