+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_push_back_v2.py
@@ -34,7 +34,7 @@ class SawyerPushBackEnvV2(SawyerXYZEnv):
         self.init_config = {
             "obj_init_pos": np.array([0, 0.8, 0.02]),
             "obj_init_angle": 0.3,
-            "hand_init_pos": np.array([0, 0.6, 0.2], dtype=np.float32),
+            "hand_init_pos": np.array([0, 0.6, 0.2], dtype=np.float64),
         }
         self.goal = np.array([0.0, 0.6, 0.02])
         self.obj_init_pos = self.init_config["obj_init_pos"]
@@ -46,8 +46,8 @@ class SawyerPushBackEnvV2(SawyerXYZEnv):
             np.hstack((obj_high, goal_high)),
         )
         self.goal_space = Box(
-            np.array(goal_low, dtype=np.float32),
-            np.array(goal_high, dtype=np.float32),
+            np.array(goal_low, dtype=np.float64),
+            np.array(goal_high, dtype=np.float64),
         )
 
     @property
